# docs/swagger.yaml
definitions:
...
  handler.ErrorBadRequest:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  handler.ErrorInternal:
    properties:
      error:
        example: some internal error
        type: string
    type: object
  handler.ErrorNotFound:
    properties:
      error:
        example: not found
        type: string
    type: object
...
  handler.ImageCreate:
    properties:
      image_url:
        type: string
      is_primary:
        type: boolean
      listing_id:
        type: integer
      order_index:
        type: integer
    type: object
  handler.ImageUpdate:
    properties:
      image_url:
        type: string
      is_primary:
        type: boolean
      order_index:
        type: integer
    type: object
...
  handler.PaymentCreate:
    properties:
      booking_id:
        type: integer
      payment_method:
        example: card
        type: string
      payment_status:
        example: completed
        type: string
    type: object
  handler.PaymentUpdate:
    properties:
      payment_method:
        type: string
      payment_status:
        type: string
    type: object
...
  handler.StatusOK:
    properties:
      message:
        example: deleted successfully
        type: string
    type: object
...
  model.Image:
    properties:
      image_id:
        type: integer
      image_url:
        type: string
      is_primary:
        type: boolean
      listing_id:
        type: integer
      order_index:
        type: integer
      uploaded_at:
        type: string
    type: object
...
  model.Payment:
    properties:
      amount:
        type: number
      booking_id:
        type: integer
      paid_at:
        type: string
      payment_id:
        type: integer
      payment_method:
        type: string
      payment_status:
        type: string
      transaction_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: DB CW
  version: "1.0"
paths:
...
  /api/bookings/{booking_id}/payments:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Payment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить платежи по booking ID
      tags:
      - payments
...
  /api/images:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные изображения
        in: body
        name: image
        required: true
        schema:
          $ref: '#/definitions/handler.ImageCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Создать изображение
      tags:
      - images
  /api/images/{id}:
    delete:
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Удалить изображение
      tags:
      - images
    get:
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить изображение по ID
      tags:
      - images
    put:
      consumes:
      - application/json
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные изображения
        in: body
        name: image
        required: true
        schema:
          $ref: '#/definitions/handler.ImageUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Обновить изображение
      tags:
      - images
...
  /api/listings/{listing_id}/images:
    get:
      parameters:
      - description: Listing ID
        in: path
        name: listing_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Image'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить изображения по listing ID
      tags:
      - images
...
  /api/payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные платежа
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/handler.PaymentCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Создать платеж
      tags:
      - payments
  /api/payments/{id}:
    delete:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Удалить платеж
      tags:
      - payments
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить платеж по ID
      tags:
      - payments
    put:
      consumes:
      - application/json
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные платежа
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/handler.PaymentUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Обновить платеж
      tags:
      - payments
...
swagger: "2.0"